Changes to flesch score

+++ MOOCTextEditor/src/document/Document.java
@@ -130,7 +130,7 @@ public abstract class Document {
 	/** return the Flesch readability score of this document */
 	public double getFleschScore()
 	{
-	    double fleschScore = (206.835-(1.015*((double) getNumWords()/getNumSentences())-84.6*((double) getNumSyllables()/getNumWords())));
+	    double fleschScore = (206.835-(1.015*((double) getNumWords()/(double) getNumSentences())-84.6*((double) getNumSyllables()/(double) getNumWords())));
 	    return fleschScore;
 	}
